refactor: text hmi
